+++ platform/darwin/src/MGLShapeSource_Private.h
@@ -16,7 +16,4 @@ namespace mbgl {
 
 @end
 
-MGL_EXPORT
-mbgl::style::GeoJSONOptions MGLGeoJSONOptionsFromDictionary(NS_DICTIONARY_OF(MGLShapeSourceOption, id) *options);
-
 NS_ASSUME_NONNULL_END