+++ src/mbgl/style/sources/geojson_source_impl.cpp
@@ -155,7 +155,7 @@ void GeoJSONSource::Impl::loadDescription(FileSource& fileSource) {
 
 Range<uint8_t> GeoJSONSource::Impl::getZoomRange() const {
     assert(loaded);
-    return { 0, options.maxzoom };
+    return { options.minzoom, options.maxzoom };
 }
 
 std::unique_ptr<Tile> GeoJSONSource::Impl::createTile(const OverscaledTileID& tileID,